Rename ota module to transfer
Rename src/ota/ to src/transfer/ to better reflect the module's
purpose as a generic file transfer mechanism. Update feature flags
(ota_mqtt_data -> transfer_mqtt, ota_http_data -> transfer_http,
ota_http_reqwest -> transfer_http_reqwest), module references, and
tests accordingly. Add AFR OTA encoding support (afr_ota.rs).
