+++ Cargo.toml
@@ -55,7 +55,7 @@ base64 = { version = "0.22", optional = true }
 
 # MQTT client implementations (optional, feature-gated)
 rumqttc = { version = "0.24", optional = true }
-greengrass-ipc-rust = { git = "https://github.com/FactbirdHQ/greengrass-ipc-rust", rev = "0f0469e", optional = true }
+greengrass-ipc-rust = { git = "https://github.com/FactbirdHQ/greengrass-ipc-rust", rev = "bcc6df9", optional = true }
 bytes = { version = "1", optional = true }
 futures = { version = "0.3", optional = true }
 reqwest = { version = "0.12", default-features = false, features = ["rustls-tls"], optional = true }
@@ -104,16 +104,16 @@ reqwest = { version = "0.12", default-features = false, features = ["rustls-tls"
 
 
 [features]
-default = ["ota_mqtt_data", "metric_cbor", "provision_cbor", "shadows_kv_persist", "mqtt_mqttrust"]
+default = ["transfer_mqtt", "metric_cbor", "provision_cbor", "shadows_kv_persist", "mqtt_mqttrust"]
 
 metric_cbor = ["dep:minicbor", "dep:minicbor-serde"]
 
 provision_cbor = ["dep:minicbor", "dep:minicbor-serde"]
 
-ota_mqtt_data = ["dep:minicbor", "dep:minicbor-serde"]
+transfer_mqtt = ["dep:minicbor", "dep:minicbor-serde"]
 
-ota_http_data = []
-ota_http_reqwest = ["ota_http_data", "std", "dep:reqwest"]
+transfer_http = []
+transfer_http_reqwest = ["transfer_http", "std", "dep:reqwest"]
 
 std = ["serde/std", "minicbor-serde?/std", "dep:tokio", "dep:serde_json", "dep:base64", "postcard/alloc", "rustot-derive/std", "embassy-time/std", "embassy-time/generic-queue-8"]
