+++ src/transfer/control_interface/mod.rs
@@ -1,9 +1,9 @@
 use crate::jobs::data_types::JobStatus;
-use crate::ota::status_details::StatusDetailsExt;
+use crate::transfer::status_details::StatusDetailsExt;
 
 use super::{
-    encoding::{json::JobStatusReason, OtaJobContext},
-    error::OtaError,
+    encoding::{json::JobStatusReason, JobContext},
+    error::TransferError,
     ProgressState,
 };
 
@@ -11,12 +11,12 @@ pub mod mqtt;
 
 // Interfaces required for OTA
 pub trait ControlInterface {
-    async fn request_job(&self) -> Result<(), OtaError>;
+    async fn request_job(&self) -> Result<(), TransferError>;
     async fn update_job_status<E: StatusDetailsExt>(
         &self,
-        job: &OtaJobContext<'_, E>,
+        job: &JobContext<'_, E>,
         progress: &mut ProgressState<E>,
         status: JobStatus,
         reason: JobStatusReason,
-    ) -> Result<(), OtaError>;
+    ) -> Result<(), TransferError>;
 }
